# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WhatsApp Clone API",
        "description": "API documentation for WhatsApp Clone - Real-time messaging application",
        "contact": {
            "email": "support@whatsappclone.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/messages/{message}/reactions": {
            "get": {
                "tags": [
                    "Reactions"
                ],
                "summary": "Get reactions for a message",
                "description": "Get reactions for a message.",
                "operationId": "2e4f5a7771eab912778adf26ce913b30",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "description": "Message ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reactions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReactionResponse"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - User is not a participant of the chat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reactions"
                ],
                "summary": "Add a reaction to a message",
                "description": "Add a reaction to a message.",
                "operationId": "9688944b0fe567383ba614666f523b94",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "description": "Message ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReactionCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reaction added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ReactionResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - User is not a participant of the chat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - Invalid emoji or duplicate reaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/{message}/reactions/{reaction}": {
            "put": {
                "tags": [
                    "Reactions"
                ],
                "summary": "Update a reaction",
                "description": "Update a reaction.",
                "operationId": "f06bbf36bfe2f24262457a826bdbb0c5",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "description": "Message ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "reaction",
                        "in": "path",
                        "description": "Reaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReactionUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reaction updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ReactionResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - User is not a participant of the chat or doesn't own the reaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - Invalid emoji",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reactions"
                ],
                "summary": "Remove a reaction",
                "description": "Remove a reaction.",
                "operationId": "a695f471dac00763d8a32f122ee1fc38",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "description": "Message ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "reaction",
                        "in": "path",
                        "description": "Reaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reaction removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reaction removed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - User is not a participant of the chat or doesn't own the reaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Validation error"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": [
                            {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "property": null
                                },
                                "property": null
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "ReactionCreateRequest": {
                "required": [
                    "emoji"
                ],
                "properties": {
                    "emoji": {
                        "description": "The emoji reaction to add",
                        "type": "string",
                        "maxLength": 10,
                        "example": "👍"
                    }
                },
                "type": "object"
            },
            "ReactionUpdateRequest": {
                "required": [
                    "emoji"
                ],
                "properties": {
                    "emoji": {
                        "description": "The new emoji reaction",
                        "type": "string",
                        "maxLength": 10,
                        "example": "❤️"
                    }
                },
                "type": "object"
            },
            "ReactionResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "message_id": {
                        "type": "integer",
                        "example": 123
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 456
                    },
                    "emoji": {
                        "type": "string",
                        "example": "👍"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 456
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            },
                            "avatar": {
                                "type": "string",
                                "example": "https://example.com/avatar.jpg"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}